refactor: Redis config
