+++ src/main/java/io/hhplus/conbook/config/RedisConfig.java
@@ -17,6 +17,8 @@ public class RedisConfig {
         // key와 value 모두 String으로 직렬화
         template.setKeySerializer(new StringRedisSerializer());
         template.setValueSerializer(new StringRedisSerializer());
+        template.setHashKeySerializer(new StringRedisSerializer());
+        template.setHashValueSerializer(new StringRedisSerializer());
 
         return template;
     }
